+++ visualization/app/codeCharta/state/store/appSettings/labelSize/labelSize.actions.ts
@@ -1,0 +1,3 @@
+import { createAction, props } from "@ngrx/store"
+
+export const setLabelSize = createAction("SET_LABEL_SIZE", props<{ value: number }>())