+++ visualization/app/codeCharta/state/store/appSettings/labelSize/labelSize.selector.ts
@@ -1,0 +1,4 @@
+import { createSelector } from "@ngrx/store"
+import { appSettingsSelector } from "../appSettings.selector"
+
+export const labelSizeSelector = createSelector(appSettingsSelector, appSettings => appSettings.labelSize)